# internal/transport/http/httpgen/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package httpgen

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn1AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn2AllowedHeaders = map[string]string{
		"PATCH": "Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/departments/"

			if l := len("/departments/"); len(elem) >= l && elem[0:l] == "/departments/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				switch r.Method {
				case "POST":
					s.handleDepartmentsPostRequest([0]string{}, elemIsEscaped, w, r)
				default:
					s.notAllowed(w, r, notAllowedParams{
						allowedMethods: "POST",
						allowedHeaders: rn1AllowedHeaders,
						acceptPost:     "application/json",
						acceptPatch:    "",
					})
				}

				return
			}
			// Param: "id"
			// Match until "/"
			idx := strings.IndexByte(elem, '/')
			if idx < 0 {
				idx = len(elem)
			}
			args[0] = elem[:idx]
			elem = elem[idx:]

			if len(elem) == 0 {
				switch r.Method {
				case "DELETE":
					s.handleDepartmentsIDDeleteRequest([1]string{
						args[0],
					}, elemIsEscaped, w, r)
				case "GET":
					s.handleDepartmentsIDGetRequest([1]string{
						args[0],
					}, elemIsEscaped, w, r)
				case "PATCH":
					s.handleDepartmentsIDPatchRequest([1]string{
						args[0],
					}, elemIsEscaped, w, r)
				default:
					s.notAllowed(w, r, notAllowedParams{
						allowedMethods: "DELETE,GET,PATCH",
						allowedHeaders: rn2AllowedHeaders,
						acceptPost:     "",
						acceptPatch:    "application/json",
					})
				}

				return
			}
			switch elem[0] {
			case '/': // Prefix: "/employees/"

				if l := len("/employees/"); len(elem) >= l && elem[0:l] == "/employees/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleDepartmentsIDEmployeesPostRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn3AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/departments/"

			if l := len("/departments/"); len(elem) >= l && elem[0:l] == "/departments/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				switch method {
				case "POST":
					r.name = DepartmentsPostOperation
					r.summary = "Create a department"
					r.operationID = ""
					r.operationGroup = ""
					r.pathPattern = "/departments/"
					r.args = args
					r.count = 0
					return r, true
				default:
					return
				}
			}
			// Param: "id"
			// Match until "/"
			idx := strings.IndexByte(elem, '/')
			if idx < 0 {
				idx = len(elem)
			}
			args[0] = elem[:idx]
			elem = elem[idx:]

			if len(elem) == 0 {
				switch method {
				case "DELETE":
					r.name = DepartmentsIDDeleteOperation
					r.summary = "Delete department"
					r.operationID = ""
					r.operationGroup = ""
					r.pathPattern = "/departments/{id}"
					r.args = args
					r.count = 1
					return r, true
				case "GET":
					r.name = DepartmentsIDGetOperation
					r.summary = "Get department details (with employees and children subtree)"
					r.operationID = ""
					r.operationGroup = ""
					r.pathPattern = "/departments/{id}"
					r.args = args
					r.count = 1
					return r, true
				case "PATCH":
					r.name = DepartmentsIDPatchOperation
					r.summary = "Update department (rename and/or change parent)"
					r.operationID = ""
					r.operationGroup = ""
					r.pathPattern = "/departments/{id}"
					r.args = args
					r.count = 1
					return r, true
				default:
					return
				}
			}
			switch elem[0] {
			case '/': // Prefix: "/employees/"

				if l := len("/employees/"); len(elem) >= l && elem[0:l] == "/employees/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = DepartmentsIDEmployeesPostOperation
						r.summary = "Create an employee in a department"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/departments/{id}/employees/"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}